+++ client/CHANGELOG.md
@@ -8,7 +8,8 @@ All notable changes to this project will be documented in this file.
 
 - Added edit My Collections functionality for names ([#657](https://github.com/project-lux/lux-frontend/issues/657)).
 - Added set default functionality for My Collections ([#657](https://github.com/project-lux/lux-frontend/issues/657)).
-- Added edit My Collections functionality for classificationss ([#657](https://github.com/project-lux/lux-frontend/issues/657)).
+- Added edit My Collections functionality for classifications ([#657](https://github.com/project-lux/lux-frontend/issues/657)).
+- Added edit My Collections functionality for identifiers ([#657](https://github.com/project-lux/lux-frontend/issues/657)).
 
 ### Changed